requirements.txt
# -----------------------
# GUI
# -----------------------
customtkinter==5.3.0
Pillow==10.0.1

# -----------------------
# Networking
# -----------------------
requests==2.31.0
urllib3==2.0.5
python-dateutil==2.8.2

# -----------------------
# System & performance monitoring
# -----------------------
psutil==5.9.5

# -----------------------
# Async & concurrency
# -----------------------
# asyncio is built-in for Python >=3.7, include for older or pip resolution
asyncio; python_version < "3.7"

# -----------------------
# LLM / AI integration (local models)
# -----------------------
torch==2.1.0; python_version >= "3.10"
transformers==4.41.0; python_version >= "3.10"
accelerate==0.23.0; python_version >= "3.10"
sentencepiece==0.1.99; python_version >= "3.10"
bitsandbytes==0.41.0; python_version >= "3.10"

# -----------------------
# Optional utilities
# -----------------------
tqdm==4.66.1
numpy==1.27.5

# -----------------------
# Cyber / pentesting helpers
# -----------------------
rich==13.5.2
colorama==0.4.7

# -----------------------
# Notes:
# -----------------------
# 1. RustScan and Amass are external binaries, install manually.
# 2. Tkinter is built-in, no pip install required.
# 3. This file supports Windows and Linux, Python 3.10–3.12. Adjust versions if using older Python.